% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_merging.R
\name{agregate}
\alias{agregate}
\title{Fonction pour agreger le graphe}
\usage{
agregate(
  m_crois,
  threshold,
  kmax = 0,
  pas_pb = floor(nrow(m_crois)/50),
  verbose = TRUE,
  ...
)
}
\arguments{
\item{m_crois}{Matrice de croisement.}

\item{threshold}{threshold de confidentialite.}

\item{kmax}{Entier indiquant le nombre d'agregtion maximale a effectuer.
Si \code{kmax} vaut 0, on agrege au maximum.}

\item{pas_pb}{Entier indiquant a quelle frequence afficher la barre de progression
(pb = progress bar). Par defaut, affiche tout les 2\% environ.}

\item{verbose}{Boolean. If TRUE, progress bar is displayed.}

\item{...}{other parameters}
}
\value{
En sortie, on a la matrice de croisement apres agregation.
}
\description{
On agrege le graphe en utilisant la fonction agregate_one un
grand nombre de fois, jusqu'a ce que plus aucune agregation ne soit possible
où jusqu'a un threshold limite d'agregations.
}
