% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_merging.R
\name{agregate_one}
\alias{agregate_one}
\title{Fonction pour agreger une seule fois 2 zones de z1}
\usage{
agregate_one(m_crois, threshold, shuffle = TRUE, verbose = TRUE, ...)
}
\arguments{
\item{m_crois}{Matrice de croisement.}

\item{threshold}{Entier indiquant le threshold de confidentialite.}

\item{shuffle}{Booleen indiquant s'il faut melanger les lignes
de \code{m_crois} apres avoir fusionne ou non. Vaut TRUE par defaut.}

\item{verbose}{Boolean. If TRUE, displays messages.}

\item{...}{other parameters}
}
\value{
En sortie on a une liste de deux elements : \code{matrice} donne
la matrice de croisement avec une ligne en moins et \code{continue} indique
si l'agregation peut être poursuivie.
}
\description{
On test toutes les paires de zones de z1 jusqu'a trouver
deux zones qu'on peut fusionner. Alors on fusionne et on s'arrête la.
}
