% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_splitting.R
\name{comp_connexe_list}
\alias{comp_connexe_list}
\title{Desagreger en composantes connexes}
\usage{
comp_connexe_list(m_crois)
}
\arguments{
\item{m_crois}{Matrice de croisement.}
}
\value{
On retourne une liste de matrices de croisement etant
chacune connexe.
}
\description{
Permet de separer la matrice de croisement en plusieurs
sous-matrices de croisement connexes.
}
