% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_splitting.R
\name{decompose_m_crois}
\alias{decompose_m_crois}
\title{Decomposer une matrice de croisement en composantes connexes}
\usage{
decompose_m_crois(m_crois, taille_max = 10)
}
\arguments{
\item{m_crois}{Matrice de croisement.}

\item{taille_max}{Entier indiquant la taille a partir de laquelle
on decide de "casser" la matrice ou non. Vaut 10 par defaut.}
}
\description{
Permet de decomposer totalement une matrice de croisement, qu'elle
soit connexe ou non, en plusieurs sous-matrices de croisements
connexes "incassables" ou ayant moins de lignes que \code{taille_max}
}
