% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_search_diff.R
\name{desagregate_list}
\alias{desagregate_list}
\title{Desagreger la liste d'agregat}
\usage{
desagregate_list(list_agregat)
}
\arguments{
\item{list_agregat}{Liste de chaînes de caractere. Chaque element de la liste
est une chaîne de caracteres donnant les nomes des zones de z1 composant l'agregat.
Chaque nom est separe dans la chaîne de caractere par le symbole ".".}
}
\value{
On retourne une liste de vecteur de chaînes de caracteres.
}
\description{
Permet de creer une liste de vecteur, chaque vecteur donnant
la composition en termes de zones de z1 de l'agregat.
}
