% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_splitting.R
\name{detect_central_node}
\alias{detect_central_node}
\title{Detecter un noeud central}
\usage{
detect_central_node(m_crois)
}
\arguments{
\item{m_crois}{Matrice de croisement.}
}
\value{
En sortie on a un entier qui indique le numero
du noeud central (c'est-a-dire le numero de la ligne).
Si aucun noeud central n'a ete trouve on retourne 0.
}
\description{
Un noeud central est un noeud qui lorsqu'on le retire
rend le graphe non connexe. Cette fonction permet de
detecter un noeud parmi les noeud centraux.
}
