% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{differencierRcpp}
\alias{differencierRcpp}
\title{Recherche exhaustive des problemes de differentiation}
\usage{
differencierRcpp(
  iTailleCible,
  iSeuil,
  vNbObsTerritoire,
  vNbObsCarreaux,
  mContiguiteT,
  mContiguiteTC
)
}
\arguments{
\item{iTailleCible}{: entiers indiquant la taille des agrégats à tester}

\item{iSeuil}{: entier indiquant le seuil en dessous duquel il y a rupture du secret statistique}

\item{vNbObsTerritoire}{: vecteur d'entiers contenant le nombre d'observations pour chaque territoire}

\item{vNbObsCarreaux}{: vecteur d'entiers contenant le nombre d'observations pour chaque carreau}

\item{mContiguiteT}{: matrice d'entiers indiquant la contiguité territoire/territoire - la diagonale est supposée à zéro}

\item{mContiguiteTC}{: matrice d'entiers indiquant la contiguité territoire/carreau - les territoires sont en ligne et les carreaux en colonne}
}
\value{
liste de vecteurs dont chaque élément est une différenciation. 
         Chaque élément est un vecteur constitué 
         - des indices des territoires de l'agrégat 
         - du nombre d'observations différenciées à l'intérieur 
         - et enfin du nombre d'observations différenciées à l'extérieur
}
\description{
Fonction exposée permettant de détecter les différenciations territoire/carreau
Elle prépare les paramètres et appelle la fonction récursive explorerRcpp
}
\details{
Remarque : Il est inutile d'explorer des tailles d'agrégats supérieures à nbTerritoires/2
}
\author{
Arlindo Dos Santos / PSAR Analyse urbaine
}
