% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffman.R
\docType{package}
\name{diffman}
\alias{diffman}
\title{\code{diffman} package}
\description{
Package to manage differentiation
}
\details{
Differentiation is a technique to deduce information on small
agregates from a source of information that has been
disseminated according to two different nomenclatures. Knowing
the value of additive variables on agregates of observations
according to the nomenclatures, one can decude the value of
these variables on new agregates. If those agregates are 
smaller than a given threshold then we can considerer there
is a confidentiality break for the observations within those
agregates. It order to respect statistical secrecy for every
observations, one need to check for the possible problematic 
differentiations that a data user can compute. The objective
of this package is to provide a tool for detecting all
observations at risk of differentiation.
}
