% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_merging_test.R
\name{is_connected_m2}
\alias{is_connected_m2}
\title{Tester s'il existe un chemin, selon la méthode 2}
\usage{
is_connected_m2(i, j, m_graph, v_arete, threshold)
}
\arguments{
\item{i, j}{Entiers indiquant deux sommets différents du graphe.}

\item{m_graph}{Matrice carré d'adjacence. Si on note a_ij l'élément
de cette matrice correspondant à la ième ligne et jème colonne, alors
a_ij=0 si les sommets i et j ne sont pas connectés et sinon a_ij est un entier
qui indique la valeur de l'arête entre i et j.}

\item{v_arete}{valeur du lien entre i et j}

\item{threshold}{threshold de confidentialité.}
}
\value{
Un booléen qui vaut TRUE si les deux sommets i et j sont
connectés, selon la méthode 2, et FALSE sinon.
}
\description{
Fonction permettant de tester s'il existe un chemin
entre les noeuds i et j où chaque arête a une valeur
plus grande que le threshold de confidentialité moins
la valeur du lien entre i et j.
}
