% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_splitting.R
\name{matrix_atomize}
\alias{matrix_atomize}
\title{Completement decomposer un graphe.}
\usage{
matrix_atomize(m_crois_connexe, etat = "cassable", taille_max)
}
\arguments{
\item{m_crois_connexe}{Matrice de croisement connexe.}

\item{etat}{Caracteres parmi "cassable" ou "incassable" indiquant
l'etat de la matrice de croisement connexe.}

\item{taille_max}{Entier indiquant a partir de quelle taille de graphe
on decide de le decomposer (le "casser").}
}
\value{
En sortie on a une liste de matrices de croisement connexes soient
"incassables" soient ayant moins de lignes que \code{taille_max}.
}
\description{
Permet de decomposer un graphe, de façon iterative, jusqu'a
ce que tous les sous-graphes soient "incassables".
}
