% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_splitting.R
\name{matrix_break}
\alias{matrix_break}
\title{Casser un graphe en plusieurs sous-graphes}
\usage{
matrix_break(m_crois_connexe)
}
\arguments{
\item{m_crois_connexe}{Matrice de croisement connexe.}
}
\value{
On retourne une liste de liste. Chaque sous-liste
est composee d'une matrice \code{m} et d'un \code{etat} parmi
"cassable" ou "incassable".
}
\description{
Cette fonction permet de decomposer un graphe
connexe en plusieurs sous-graphes connnexe
apres avoir detecter un noeud central.
}
