% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_reshaping.R
\name{matrix_crois}
\alias{matrix_crois}
\title{Obtenir une matrice (sparse) qui croise les zones de z1 et celles de z2}
\usage{
matrix_crois(t_crois)
}
\arguments{
\item{t_crois}{Table de croisement (1 colonne z1 et une colonne z2).}
}
\value{
En sortie on obtient une matrice sparse de croisement. Les noms
des lignes correspondent aux noms des zones de z1 et les noms des colonnes
aux noms des zones de z2.
}
\description{
La matrice de croisement est une matrice dont les lignes indiquent les zones
du zoange z1 et les colonnes les zones du zonage z2. Chaque élément de la
matrice donne le nombre d'observations situées à l'intersection entre une
zone de z1 et une zone de z2.
}
