% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_reshaping.R
\name{matrix_crois_from_tind}
\alias{matrix_crois_from_tind}
\title{Creation de la matrice de croisement}
\usage{
matrix_crois_from_tind(t_ind)
}
\arguments{
\item{t_ind}{La table individuelle (format data.table) Chaque ligne représente une observation
et elle doit au moins contenir deux colonnes nommées 'z1' et 'z2' et indiquant
à quels zonages appartient l'observation.}
}
\value{
matrice sparse de croisement
}
\description{
Permet de creer directement la matrice
de croisement a partir de la table individuelle
des observations. Wrapper pour les fonctions
simplify_z2_rem, tab_crois, simplify_z2_fus,
matrix_crois et simplify_mcrois
}
