% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_reshaping.R
\name{matrix_graphe}
\alias{matrix_graphe}
\title{Constuire la matrice d'adjacence du graphe}
\usage{
matrix_graphe(m_crois, multi = TRUE)
}
\arguments{
\item{m_crois}{Matrice de croisement.}

\item{multi}{Booléen indiquant s'il faut considérer les zones de z2
recouvrant trois zones ou plus de z1.}
}
\value{
En sortie on obtient une matrice carré d'adjacence.
}
\description{
Fonction permettant à partir de la matrice de croisement
de déterminer la matrice d'adjacence du graphe. Cette matrice
de graphe est pondérée et non symmétrique (ce qui correspond
à un graphe orienté).
}
\details{
L'option \code{multi} permet de choisir si on prend en compte
ou non les zones de z2 recouvrant 3 zones de z1 ou plus. En effet
si on les prend en compte, alors certaines observations sont comptées
plusieurs fois dans le graphe, ce qui peut conduire à de mauvaises
interprétations.
}
