% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_reshaping.R
\name{matrix_liens}
\alias{matrix_liens}
\title{Créer la matrice de liens (ou matrice de contiguïté)}
\usage{
matrix_liens(m_crois)
}
\arguments{
\item{m_crois}{Matrice de croisement.}
}
\value{
En sortie on a une matrice carré de booléens.
}
\description{
Cette fonction permet de créer une matrice carré de booléens de taille
égale au nombre de zones du zonage z1.
}
\details{
Un élément (i,j) de cette matrice
vaut TRUE (ou 1) si les zones i et j du zonage z1 sont contigues, c'est-à-dire
s'il existe au moins une zone de z2 recouvrant à la fois i et j. Les élements
de la diagonales portent la valeur FALSE.
}
