% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_search_diff.R
\name{search_diff_agregate}
\alias{search_diff_agregate}
\title{Tester toutes les combinaisons possibles (jusqua une certaine taille)}
\usage{
search_diff_agregate(list_m_crois, threshold, max_agregate_size = 20)
}
\arguments{
\item{list_m_crois}{Liste de matrices de croisement.}

\item{threshold}{threshold de confidentialite.}

\item{max_agregate_size}{Entier indiquant la taille maximale des agregats
a tester}
}
\value{
On retourne une liste d'agregats.
}
\description{
Permet de tester, pour une liste de matrice de croisement, les agregats
de zones de z1 conduisant a un probleme de differenciation.
}
