% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_reshaping.R
\name{simplify_mcrois}
\alias{simplify_mcrois}
\title{Simplifier la matrice de croisement}
\usage{
simplify_mcrois(m_crois)
}
\arguments{
\item{m_crois}{Une matrice de croisement.}
}
\value{
En matrice on a une matrice de croisement simplifiée, c'est-à-dire
avec moins de colonnes.
}
\description{
Simplifier la matrice de croisement revient à supprimer deux types de colones
(et donc supprimer des zones du zonage z2).
1 - on supprime les colonnes vides, c'est-à-dire ne contenant que des 0.
2 - on supprime les colonnes ne contenant qu'un seul élément différent de 0.
}
