% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_reshaping.R
\name{simplify_z2_fus}
\alias{simplify_z2_fus}
\title{Fusionne les zones de z2 recouvrant les mêmes zones de z1}
\usage{
simplify_z2_fus(t_crois)
}
\arguments{
\item{t_crois}{Table de croisement (c'est la table de fréquences lorsqu'on
croise les deux zonages z1 et z2).}
}
\value{
En sortie on obtient une table de croisement avec moins de classes
pour le zonage z2.
}
\description{
Cette fonction permet de fusionner les zones de z2 dont les observations
sont réparties sur les mêmes zones de z1. Cela permet de diminuer le
nombre de zonages de z2.
}
