% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_reshaping.R
\name{simplify_z2_rem}
\alias{simplify_z2_rem}
\title{Enlever les zones de z2 entièrement incluses dans une zone de z1}
\usage{
simplify_z2_rem(t_ind)
}
\arguments{
\item{t_ind}{La table individuelle (format data.table) Chaque ligne représente une observation
et elle doit au moins contenir deux colonnes nommées 'z1' et 'z2' et indiquant
à quels zonages appartient l'observation.}
}
\value{
En sortie on récupère une table individuelle avec des observations en
moins
}
\description{
Cette fonction permet de retirer de la table individuelle, les observations
contenues dans une zone de z2 entièrement incluse dans une seule zone de z1.
Une zone de z2 est entièrement incluse dans une autre zone si toutes les
observations qu'elle contient sont également contenue dans cette autre zone.
}
