% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{t_ex}
\alias{t_ex}
\title{A simulated dataset of 34695 observations}
\format{A data frame with 34695 rows and 3 variables:
\describe{
  \item{id}{number to identify the observation}
  \item{z1}{first nomenclature. Each category is coded by an integer (between 5 and 363)}
  \item{z2}{second nomenclature. Each category is coded by an integer (between 1 and 406)}
}}
\usage{
t_ex
}
\description{
A dataset containing the categories observations belongs to.
Each observation is included in two categories : one from
the nomenclature z1 and one from the nomenclature z2.
}
\keyword{datasets}
