% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_reshaping.R
\name{tab_crois}
\alias{tab_crois}
\title{Creer la matrice de croisement entre deux zonages}
\usage{
tab_crois(t_ind)
}
\arguments{
\item{t_ind}{La table individuelle (format data.table) Chaque ligne représente une observation
et elle doit au moins contenir deux colonnes nommées 'z1' et 'z2' et indiquant
à quels zonages appartient l'observation.}
}
\value{
table de croisement
}
\description{
Compte le nombre d'observations aux intersections des 
deux zonages.
}
