% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_merging_test.R
\name{test_fus}
\alias{test_fus}
\title{Tester si on peut fusionner deux zones de z1}
\usage{
test_fus(i, j, m_crois, threshold, methode = "both")
}
\arguments{
\item{i, j}{Entiers indiquant les zones de z1.}

\item{m_crois}{Matrice de croisement.}

\item{threshold}{threshold de confidentialité.}

\item{methode}{Méthode à choisir parmi "m1", "m2" et "both".}
}
\value{
En sortie, on a une liste contenant deux élements, fus et col.
fus vaut TRUE ou FALSE et indique s'il faut fusionner les deux lignes.
col est un vecteur d'entier qui indique les numéros de colonne à supprimer.
}
\description{
On teste selon la méthode 1, 2 ou les deux à la fois, si
deux zones i et j de z1 peuvent être fusionnées ou non
}
