% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_merging_test.R
\name{test_fus_m2}
\alias{test_fus_m2}
\title{Tester, selon la méthode 2, si on peut fusionner deux noeuds ou non.}
\usage{
test_fus_m2(i, j, m_crois, threshold)
}
\arguments{
\item{i, j}{Entiers indiquant les noeuds du graphe à tester.}

\item{m_crois}{Matrice de croisement.}

\item{threshold}{threshold de confidentialité.}
}
\value{
En sortie, on a une liste contenant deux élements, fus et col.
fus vaut TRUE ou FALSE et indique s'il faut fusionner les deux lignes.
col est un vecteur d'entier qui indique les numéros de colonne à supprimer.
}
\description{
Permet de tester si les deux sommets i et j du graphe peuvent
être fusionnés.
}
\details{
La méthode 2 s'applique après la méthode 1. Si les sommets i et j
sont reliés par une arête en-dessous du threshold de confidentialité
la méthode 1 conclut qu'on ne peut pas les fusionner. Mais la méthode
2 regarde s'il existe un autre chemin entre i et j, avec des arêtes
ayant des valeurs suffisamment élevées (plus grande que le threshold - la valeur
du lien entre i et j). Si c'est le cas, on peut en déduire qu'il n'y
aura pas de problème de différentiation en considérant i et j séparément.
On peut donc fusionner ces deux zones.
}
