% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_merging_test.R
\name{valeurs_lien}
\alias{valeurs_lien}
\title{Valeurs des arêtes reliant i et j}
\usage{
valeurs_lien(i, j, m_crois, ...)
}
\arguments{
\item{i, j}{Entiers indiquant les zones de z1.}

\item{m_crois}{Matrice de croisement.}

\item{...}{d'autres paramètres hérités de \code{zones_front}.}
}
\value{
En sortie on obtient un vecteur de deux entiers
donnant la valeur de l'arête de i vers j et de celle de j
vers i.
}
\description{
i et j étant deux zones de z1, cette fonction donne la valeur
de l'arête reliant i et j et de celle reliant j à i.
}
