% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_merging_test.R
\name{zones_front}
\alias{zones_front}
\title{Détermine les zones de z2 reouvrant plusieurs zones de z1}
\usage{
zones_front(i, j, m_crois, type = "all")
}
\arguments{
\item{i, j}{Entiers indiquant les zones de z1.}

\item{m_crois}{Matrice de croisement.}

\item{type}{Character entre "all", "multi" et "unique", pour
sélectionner le type de zones de z2 à prendre en compte.}
}
\value{
En sortie on a un vecteur d'entiers indiquant les numéros
de colonnes de \code{m_crois} impliqués.
}
\description{
Cette fonction permet de déterminer les colonnes de \code{m_crois}
correspondant aux zones de z2 à la frontière entre les zones i et j
de z1.
}
